test: centralize emulated account lookup
