+++ apps/web/__tests__/playwright/emulated/chat/chat-test-helpers.ts
@@ -2,38 +2,6 @@ import type { Page } from "@playwright/test";
 import { Client } from "pg";
 
 export const SEEDED_CHAT_ID = "playwright-manage-chat";
-const ACCOUNT_LOOKUP_TIMEOUT_MS = 60_000;
-
-export async function getEmailAccountId(page: Page) {
-  let lastError: unknown;
-  const deadline = Date.now() + ACCOUNT_LOOKUP_TIMEOUT_MS;
-
-  while (Date.now() < deadline) {
-    try {
-      const response = await page.request.get("/api/user/email-accounts");
-      if (response.ok()) {
-        const { emailAccounts } = (await response.json()) as {
-          emailAccounts: { id: string }[];
-        };
-        const emailAccountId = emailAccounts[0]?.id;
-        if (emailAccountId) return emailAccountId;
-        lastError = new Error("The setup project created no account");
-      } else {
-        lastError = new Error(
-          `Email accounts request failed with ${response.status()}`,
-        );
-      }
-    } catch (error) {
-      lastError = error;
-    }
-
-    await page.waitForTimeout(1000);
-  }
-
-  throw new Error(
-    `Could not read the setup email account: ${getErrorMessage(lastError)}`,
-  );
-}
 
 export async function markAssistantOnboardingViewed(page: Page) {
   await page.goto("/");
@@ -111,8 +79,4 @@ export async function getChatState(emailAccountId: string) {
 
 async function deleteSeededChat(client: Client) {
   await client.query(`DELETE FROM "Chat" WHERE id = $1`, [SEEDED_CHAT_ID]);
-}
-
-function getErrorMessage(error: unknown) {
-  return error instanceof Error ? error.message : String(error);
 }
