+++ apps/web/__tests__/playwright/emulated/mail/mail-test-helpers.ts
@@ -1,4 +1,5 @@
 import { expect, type Locator, type Page } from "@playwright/test";
+import { getEmailAccountId } from "../account-test-helpers";
 
 export async function openMail(page: Page) {
   const emailAccountId = await getEmailAccountId(page);
@@ -18,15 +19,4 @@ export function conversationWithSubject(
   return conversations
     .getByRole("option")
     .filter({ has: page.getByText(subject, { exact: true }) });
-}
-
-async function getEmailAccountId(page: Page) {
-  const response = await page.request.get("/api/user/email-accounts");
-  expect(response.ok()).toBeTruthy();
-  const { emailAccounts } = (await response.json()) as {
-    emailAccounts: { id: string }[];
-  };
-  const emailAccountId = emailAccounts[0]?.id;
-  if (!emailAccountId) throw new Error("The setup project created no account");
-  return emailAccountId;
 }
